REFACTOR: 주소 필드를 roadAddress로 업데이트하고 지역 유형을 BigDecimal로 변경

+++ src/main/java/cloud/intensive/kpt/domain/apartment/dto/UnitInfoRes.java
@@ -2,6 +2,8 @@ package cloud.intensive.kpt.domain.apartment.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 
+import java.math.BigDecimal;
+
 public record UnitInfoRes(
 
         @Schema(example = "1")
@@ -17,5 +19,5 @@ public record UnitInfoRes(
         String unitNumber,
 
         @Schema(description = "전용면적(㎡)", example = "84")
-        Integer area
+        BigDecimal area
 ) {}
